+++ Src/Roflcopter.Plugin.Tests/test/data/MismatchedFileNames/PartialClass.cs
@@ -1,0 +1,6 @@
+﻿namespace Roflcopter.Sample.MismatchedFileNames
+{
+    public partial class PartialClass
+    {
+    }
+}