+++ Src/Roflcopter.Plugin.Tests/test/data/MismatchedFileNames/QuickFix/SomeClass.cs
@@ -1,0 +1,6 @@
+﻿namespace Roflcopter.Sample.MismatchedFileNames.QuickFix
+{
+    public class SomeClass
+    {
+    }
+}